rename: 유틸함수, 유효성 함수 파일 위치 변경
